+++ src/__tests__/App.test.jsx
@@ -1,13 +1,23 @@
-import { describe, expect, it } from "vitest";
-import { render, screen } from "@testing-library/react";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import {
+  render,
+  screen,
+  waitForElementToBeRemoved,
+} from "@testing-library/react";
 import { createBrowserRouter } from "react-router-dom";
 import App from "../App";
 import routes from "../routes/routes";
 
+beforeEach(() => {
+  window.scrollTo = vi.fn();
+});
+
 describe("App component", () => {
-  it("The App component matches snapshot", () => {
-    const router = createBrowserRouter(routes, { initialEntries: ["/"] });
+  it("The App component matches snapshot", async () => {
+    const router = createBrowserRouter(routes);
     const { container } = render(<App router={router} />);
+
+    await waitForElementToBeRemoved(() => screen.getByTitle("Loading"));
     expect(container).toMatchSnapshot();
   });