+++ src/__tests__/Shop.test.jsx
@@ -1,7 +1,12 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { render, screen } from "@testing-library/react";
+import {
+  render,
+  screen,
+  waitForElementToBeRemoved,
+} from "@testing-library/react";
 import { RouterProvider, createMemoryRouter } from "react-router-dom";
 import { fetchMock } from "./mocks";
+import App from "../App";
 import { CartProvider } from "../components/Cart";
 import routes from "../routes/routes";
 
@@ -24,12 +29,16 @@ beforeEach(() => {
 
 describe("Shop component", () => {
   it("The Shop component component matches snapshot", async () => {
-    const router = createMemoryRouter(routes, { initialEntries: ["/shop"] });
-    const { container } = render(
-      <CartProvider>
-        <RouterProvider router={router} />
-      </CartProvider>
-    );
+    // False positive
+    const router = createMemoryRouter(routes, {
+      initialEntries: ["/shop"],
+    });
+
+    const { container } = render(<App router={router} />);
+    await waitForElementToBeRemoved(() => screen.queryByTitle("Loading"), {
+      timeout: 2000,
+    });
+
     expect(container).toMatchSnapshot();
   });
 
@@ -38,25 +47,30 @@ describe("Shop component", () => {
     window.fetch.mockRejectedValue(
       new Error("API is down", { cause: { ...rejectCause } })
     );
-    const router = createMemoryRouter(routes, { initialEntries: ["/shop"] });
 
-    render(
-      <CartProvider>
-        <RouterProvider router={router} />
-      </CartProvider>
-    );
+    const router = createMemoryRouter(routes, {
+      initialEntries: ["/shop"],
+    });
+
+    render(<App router={router} />);
+    await waitForElementToBeRemoved(() => screen.queryByTitle("Loading"), {
+      timeout: 2000,
+    });
 
     const errorMessage = await screen.findByText("API is down");
+
     expect(errorMessage).toBeInTheDocument();
   });
 
   it("The 'All' category is first in the categories list", async () => {
-    const router = createMemoryRouter(routes, { initialEntries: ["/shop"] });
-    render(
-      <CartProvider>
-        <RouterProvider router={router} />
-      </CartProvider>
-    );
+    const router = createMemoryRouter(routes, {
+      initialEntries: ["/shop"],
+    });
+
+    render(<App router={router} />);
+    await waitForElementToBeRemoved(() => screen.queryByTitle("Loading"), {
+      timeout: 2000,
+    });
 
     const lists = await screen.findAllByRole("list");
     const categoriesListItems = lists.find((list) =>
@@ -64,6 +78,7 @@ describe("Shop component", () => {
     ).children;
 
     const firstCategory = categoriesListItems[0];
+
     expect(firstCategory.textContent).match(/all/i);
   });
 });